Move schedule impl to system module
